Make sendSlackNotif function a mixin
